[finish #162004524]finished working on most endpoints
