+++ .env
@@ -1,7 +1,8 @@
+source venv/bin/activate
 export FLASK_ENV=development
 export FLASK_DEBUG=True
 export FLASK_APP=run.py
 export DATABASE_DEVELOPMENT="postgresql://postgres:new1@localhost:5432/SendIT"
-export DATABASE_TEST="postgresql://postgres:new1@localhost:5432/sendit_test"
+export DATABASE_TEST="postgresql://postgres:new1@localhost:5432/test_db"
 export DATABASE_PRODUCTION="postgresql://postgres:new1@localhost:5432/SendIT_production"
