+++ app/tests/test_user.py
@@ -4,16 +4,20 @@ import json
 import sys
 import os
 
-sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 from app.tests.base_test import BaseTest
 
 
 class TestUser(BaseTest):
 
-    def test_create_user(self):
-        resp = self.client.post('/api/v1/users', data=json.dumps(self.test_create_user), content_type='application/json')
-        self.assertEqual(resp.status_code, 200)
+    def test_register_user(self):
+        resp = self.client.post('/api/v2/auth/signup', json=self.register_user)
+        self.assertEqual(resp.status_code, 201)
+
+    def test_inexistent_user(self):
+        resp = self.client.post('/api/v2/auth/login', json=self.login_user)
+        self.assertEqual(resp.status_code, 400)
+
 
 
 if __name__ == '__main__':
-    unittest.main()
+     unittest.main()
